% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimateSummary}
\alias{estimateSummary}
\title{Summarize parameter estimates from stan fit.}
\usage{
estimateSummary(posterior, params, predType)
}
\arguments{
\item{posterior}{Posterior distribution of stan fit, created using rstan::extract(stanFitObject)}

\item{params}{Character vector of parameter names for those parameters estimates to be summarized. Options can be queried using `names(posterior)`. `params` can only contain the name(s) of one of the main parameters estimated per trial (e.g., not "sigma") or transformed parameters estimated per trial (e.g., not "predicted_concentration" or "lp__", see predConcSumm() to summarize predicted concentration).}

\item{predType}{Single-item character vector with the type of prediction used in the stan fit (typically either "power law RTD" or "exponential RTD")}
}
\description{
Returns a dataframe containing the mean, median sd and se of the parameter estimates per trial, summarized across iterations of the stan model fit for all parameters included in `params`.
}
