% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_stan_data_reqs}
\alias{get_stan_data_reqs}
\title{Query data{} block requirements for a given stan file}
\usage{
get_stan_data_reqs(stan_file_name, file_path = NULL)
}
\arguments{
\item{stan_file_name}{Single item character vector that defines the file path + file name + .stan extension of the stan file to query data structure for}

\item{file_path}{Default is NULL, which applies when the stan file is saved in the working directory. Otherwise, specify the single-item character vector defining the file path that contains the stan file}
}
\description{
Query data{} block requirements for a given stan file
}
