% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_matrix}
\alias{make_matrix}
\title{Make 3D matrix of channel concentration values.}
\usage{
make_matrix(results_tbl, target = "C_c")
}
\arguments{
\item{target}{A character vector of length 1, default is "C_c" (channel concentration)}

\item{result_tbl}{A tibble of concentration values relative to the three dimensions of the eventual matrix (created using hydrogeom and flumeTracer)}
}
\value{
3D concentration matrix with dimensionless concentration values 
(decaying from 1 to 0, e.g., scaled relative to initial concentration and 
final concentration); and relative to dimensionless time values 
(e.g., increasing from 0 to 1 as a fraction of maximum residence time, 
time 0 is when "C_c" = 1.0, time 1.0 is when "C_c" = 0)
}
\description{
Make 3D concentration matrix relative to the maximum residence time (tau_n), 
the ratio of volume between the hyporheic zone : total (V_ratio) 
and the curvature parameter in the residence time distribution 
(alpha for power law or sigma for exponential functional forms)
}
