% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{paramPostDists}
\alias{paramPostDists}
\title{Extract posterior distributions for parameter estimates}
\usage{
paramPostDists(posterior, param, dataDF = NA)
}
\arguments{
\item{posterior}{is the posterior distributions from rstan::extract() of the stanFit object}

\item{param}{is the desired parameter to extract (e.g., "tau_n", "predicted_concentration"). If param = "predicted_concentration", dataDF cannot be NA}

\item{dataDF}{(only required if `param` = "predicted_concentration"; default=NA) is the original dataset dataframe fit by stan (e.g., nrow(dataDF) must be equal to the N in the list of data sent to the stan model), which must include columns `trialIdx` identifying the trials (with values from 1:trialN), `time` and `C` (concentration). The default is NA, as this is not required when param != "predicted_concentration"}
}
\description{
Extract posterior distributions for parameter estimates
}
