% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{predConcSummary}
\alias{predConcSummary}
\title{Summary of predicted concentration lines across iterations.}
\usage{
predConcSummary(posterior, dataDF, summary, summaryArgs, predType)
}
\arguments{
\item{posterior}{Posterior distribution of stan fit (result of rstan::extract(stanFitObject)).}

\item{dataDF}{Original dataset dataframe fit by stan (e.g., nrow(dataDF) must be equal to the N in the list of data sent to the stan model), which must include columns `trialIdx` identifying the trials (with values from 1:trialN), `time` and `C` (concentration)}

\item{summary}{Desired summary type. Either "quantile" (based on quantiles of predicted concentration), or "sample" (random sample of iterations)}

\item{summaryArgs}{If `summary`="quantile", a two-item vector of upper and lower quantile limits to summarize predicted concentration across iterations (e.g., c(0.0275, 0.975)). If `summary`="sample", a single-item numeric vector defining the number of iterations to sample in the predicted concentration summary table.}

\item{predType}{Single item character vector defining the prediction type, typically either "exponential RTD" or "power law RTD"}
}
\description{
Returns list with two dataframes, `predSummary` and `iterRMSE`. `predSummary` contains a summary of the predicted concentrations across iterations, either based on median/quantile of iteration predictions or based on a random sample of predicted concentration iterations.
}
