% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dedensify}
\alias{dedensify}
\alias{dedensify_df}
\title{Dedensify data using a log10 spacing}
\usage{
dedensify(x, y, n, x_offset = 1, decimals = 10)

dedensify_df(x, x_col, y_col, n, ...)
}
\arguments{
\item{x}{vector of data values in x dimension or a data.frame}

\item{y}{vector of data values in y dimension (equal in length to x)}

\item{n}{number of samples to extract}

\item{x_offset}{value added to x before log10 is applied; large values
provide more evenly spaced points.  Small values crowd points toward the
left (smaller values of x).}

\item{x_col, ycol}{name or number of data.frame column containing x dimension
and y dimension}

\item{...}{other values passed on to dedensify.}
}
\value{
a tibble of n {x,y} pairs
}
\description{
Sample data using evenly spaced orders of magnitude.
}
