% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dedensify}
\alias{dedensify}
\title{Dedensify data using a log10 spacing}
\usage{
dedensify(x, y, n)
}
\arguments{
\item{x}{vector of data values in x dimension}

\item{y}{vector of data values in y dimension (equal in length to x)}

\item{n}{number of samples to extract}
}
\value{
a tibble of n {x,y} pairs
}
\description{
Sample data using evenly spaced orders of magnitude.
}
